requirements.txt
aws-cdk-lib<3.0.0,>=2.50.0
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.core==2.27.0
aws-solutions-constructs.aws-cloudfront-s3==2.27.0